scripts: give l2DB parser scripts more decent filename
